third_party/WebKit/LayoutTests/fast/multicol/client-rects-sole-empty-block.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>body { margin:8px; }</style>
<div style="margin:100px; columns:2; column-gap:20px; width:420px;">
    <div id="empty"></div>
</div>
<script>
    test(function() {
        var rects = document.getElementById("empty").getClientRects();
        assert_equals(rects.length, 1);
        assert_equals(rects[0].left, 108);
        assert_equals(rects[0].top, 100);
        assert_equals(rects[0].right, 308);
        assert_equals(rects[0].bottom, 100);
    }, "Zero-height block in otherwise empty multicol");
</script>
